+++ packages/core-js/modules/esnext.bigint.range.js
@@ -1,6 +1,6 @@
 'use strict';
 var $ = require('../internals/export');
-var RangeIterator = require('../internals/range-iterator');
+var NumericRangeIterator = require('../internals/numeric-range-iterator');
 
 // `BigInt.range` method
 // https://github.com/tc39/proposal-Number.range
@@ -8,7 +8,7 @@ if (typeof BigInt == 'function') {
   $({ target: 'BigInt', stat: true }, {
     range: function range(start, end, option) {
       /* global BigInt -- safe */
-      return new RangeIterator(start, end, option, 'bigint', BigInt(0), BigInt(1));
+      return new NumericRangeIterator(start, end, option, 'bigint', BigInt(0), BigInt(1));
     }
   });
 }