+++ packages/core-js/modules/esnext.number.range.js
@@ -1,11 +1,11 @@
 'use strict';
 var $ = require('../internals/export');
-var RangeIterator = require('../internals/range-iterator');
+var NumericRangeIterator = require('../internals/numeric-range-iterator');
 
 // `Number.range` method
 // https://github.com/tc39/proposal-Number.range
 $({ target: 'Number', stat: true }, {
   range: function range(start, end, option) {
-    return new RangeIterator(start, end, option, 'number', 0, 1);
+    return new NumericRangeIterator(start, end, option, 'number', 0, 1);
   }
 });